requirements.txt
# 阿贝云免费服务器自动延期程序依赖
# 创建日期: 2026-02-17

# HTTP请求库
requests>=2.31.0

# 定时任务调度
APScheduler>=3.10.4

# 日志处理
python-json-logger>=2.0.7

# 配置管理
python-dotenv>=1.0.0

# 邮件通知（可选）
# yagmail>=0.15.293
